feat: input forwarding and event handling for zed_web browser client
- Connection: typed action-forwarding methods (click, scroll, key, hover,
  mouse down/up, viewport changed, request snapshot, pong)
- RemoteView: attach click/scroll handlers on interactive DisplayNodes,
  forward events to server via Connection
- WebWindow: full GPUI input pipeline integration — mouse move/down/up,
  scroll wheel, key down/up dispatched as PlatformInput events; canvas
  made focusable for keyboard capture
- Launch: sends initial viewport size, installs browser resize and
  keyboard listeners that forward to server
- Proto: protobuf wire format for DisplayTree (display_tree.proto)
- Streaming platform: Linux dispatcher for headless-web mode
- GPUI: headless-web feature flag, display_tree module, capture hooks

+++ crates/gpui/src/text_system/line.rs
@@ -76,7 +76,7 @@ impl ShapedLine {
             if let Some(builder) = window.display_tree_builder.as_mut() {
                 let bounds = Bounds::new(origin, size(self.layout.width, line_height));
                 let runs = self.decoration_runs.iter().map(|r| DisplayTextRun {
-                    len: r.len as usize,
+                    len: r.len as u64,
                     color: Some(DisplayColor::from_hsla(r.color)),
                     font_weight: None,
                     italic: false,
@@ -167,7 +167,7 @@ impl WrappedLine {
                 let height = line_height * (self.layout.wrap_boundaries.len() as f32 + 1.);
                 let bounds = Bounds::new(origin, size(self.layout.unwrapped_layout.width, height));
                 let runs = self.decoration_runs.iter().map(|r| DisplayTextRun {
-                    len: r.len as usize,
+                    len: r.len as u64,
                     color: Some(DisplayColor::from_hsla(r.color)),
                     font_weight: None,
                     italic: false,
